# BACKEND/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Application",
        "description": "Documentation complète de l'API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/missions": {
            "get": {
                "tags": [
                    "Missions"
                ],
                "summary": "Lister toutes les missions",
                "description": "Retourne la liste de toutes les missions triées par date de début et libellé",
                "operationId": "8535defd7bbaba602b073824e4b7bc49",
                "responses": {
                    "200": {
                        "description": "Liste des missions récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Mission"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Missions"
                ],
                "summary": "Créer une nouvelle mission",
                "description": "Crée une nouvelle mission avec les informations fournies",
                "operationId": "331c4fedf7fcde3d350d6f9c256d583f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "description",
                                    "from"
                                ],
                                "properties": {
                                    "label": {
                                        "description": "Libellé de la mission",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Mission de livraison urgente"
                                    },
                                    "description": {
                                        "description": "Description détaillée de la mission",
                                        "type": "string",
                                        "example": "Livraison de matériel médical vers l'hôpital central"
                                    },
                                    "from": {
                                        "description": "Date de début de la mission",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "to": {
                                        "description": "Date de fin de la mission (optionnelle)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-20",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mission créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Mission"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/missions/{id}": {
            "get": {
                "tags": [
                    "Missions"
                ],
                "summary": "Afficher une mission",
                "description": "Retourne les détails d'une mission spécifique avec ses véhicules, utilisateurs et chauffeurs associés",
                "operationId": "e3e0a43094ae1c0093e0ce6cabd49c08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la mission",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mission récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MissionWithRelations"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mission non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Missions"
                ],
                "summary": "Mettre à jour une mission",
                "description": "Met à jour les informations d'une mission existante",
                "operationId": "f3aedd3854cadb68bc96ca70a7af8668",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la mission",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "description",
                                    "from"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Mission de livraison urgente - Mise à jour"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Livraison de matériel médical vers l'hôpital central - Priorité haute"
                                    },
                                    "from": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-16"
                                    },
                                    "to": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-22",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mission mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MissionWithRelations"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mission non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Missions"
                ],
                "summary": "Supprimer une mission",
                "description": "Supprime une mission si elle n'a pas de réservations associées",
                "operationId": "40a92f27a9ab00b081855c64d07ef2e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID de la mission",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mission supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Impossible de supprimer (réservations existantes)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mission non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Opération réussie"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Une erreur est survenue"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "Image": {
                "title": "Image",
                "description": "Schéma de données d'une image",
                "required": [
                    "path"
                ],
                "properties": {
                    "path": {
                        "description": "Chemin d'accès à l'image",
                        "type": "string",
                        "example": "path/to/file.jpg"
                    }
                },
                "type": "object"
            },
            "Mission": {
                "description": "Schéma de base d'une mission",
                "required": [
                    "id",
                    "label",
                    "description",
                    "from"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "label": {
                        "type": "string",
                        "example": "Mission de livraison urgente"
                    },
                    "description": {
                        "type": "string",
                        "example": "Livraison de matériel médical"
                    },
                    "from": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "to": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-20",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "MissionWithRelations": {
                "description": "Mission avec toutes ses relations",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Mission"
                    },
                    {
                        "properties": {
                            "vehicles": {
                                "description": "Véhicules affectés à la mission",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Vehicle"
                                },
                                "nullable": true
                            },
                            "users": {
                                "description": "Utilisateurs affectés à la mission",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/User"
                                },
                                "nullable": true
                            },
                            "drivers": {
                                "description": "Chauffeurs affectés à la mission",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/User"
                                },
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "User": {
                "description": "Schéma de données d'un utilisateur",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440002"
                    },
                    "firstName": {
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "user"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+33123456789",
                        "nullable": true
                    },
                    "isActive": {
                        "type": "boolean",
                        "example": true
                    },
                    "image": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Image"
                            }
                        ],
                        "nullable": true
                    },
                    "token": {
                        "type": "string",
                        "example": "1|abcdefgh1234567890",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Vehicle": {
                "description": "Schéma de données d'un véhicule",
                "required": [
                    "id",
                    "brand",
                    "model",
                    "type",
                    "registrationNumber",
                    "registrationDate",
                    "seatsCount",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "brand": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "type": "string",
                        "example": "Corolla"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "car",
                            "motorcycle",
                            "truck",
                            "van",
                            "bus"
                        ],
                        "example": "car"
                    },
                    "registrationNumber": {
                        "type": "string",
                        "example": "AB-123-CD"
                    },
                    "registrationDate": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-15"
                    },
                    "seatsCount": {
                        "type": "integer",
                        "example": 5
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "reserved",
                            "suspended",
                            "under_repair"
                        ],
                        "example": "available"
                    },
                    "reason": {
                        "type": "string",
                        "example": "En révision technique",
                        "nullable": true
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Image"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification par token Bearer",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}